refactor(toolset): extract Catalog into catalog.rs
Move Catalog, CatalogEntry, and helper functions (search scoring,
normalization) out of mod.rs into a dedicated catalog.rs file.
mod.rs retains only ToolSets (the top-level service).
